+++ src/Controller/AdminControllerTrait.php
@@ -16,6 +16,7 @@ use EasyCorp\Bundle\EasyAdminBundle\Form\Type\EasyAdminBatchFormType;
 use EasyCorp\Bundle\EasyAdminBundle\Form\Type\EasyAdminFiltersFormType;
 use EasyCorp\Bundle\EasyAdminBundle\Form\Type\EasyAdminFormType;
 use EasyCorp\Bundle\EasyAdminBundle\Search\Paginator;
+use EasyCorp\Bundle\EasyAdminBundle\Search\QueryPaginator;
 use Symfony\Component\EventDispatcher\GenericEvent;
 use Symfony\Component\Form\Extension\Core\Type\HiddenType;
 use Symfony\Component\Form\Extension\Core\Type\SubmitType;
@@ -615,7 +616,7 @@ trait AdminControllerTrait
      * @param string|null $sortDirection
      * @param string|null $dqlFilter
      *
-     * @return Paginator The paginated query results
+     * @return QueryPaginator The paginated query results
      */
     protected function findAll($entityClass, $page = 1, $maxPerPage = 15, $sortField = null, $sortDirection = null, $dqlFilter = null)
     {
@@ -633,7 +634,7 @@ trait AdminControllerTrait
             'sort_direction' => $sortDirection,
         ]);
 
-        return $this->get('easyadmin.paginator')->create($queryBuilder, $page, $maxPerPage);
+        return new QueryPaginator($queryBuilder, $page, $maxPerPage);
     }
 
     /**
@@ -664,7 +665,7 @@ trait AdminControllerTrait
      * @param string|null $sortDirection
      * @param string|null $dqlFilter
      *
-     * @return Paginator The paginated query results
+     * @return QueryPaginator The paginated query results
      */
     protected function findBy($entityClass, $searchQuery, array $searchableFields, $page = 1, $maxPerPage = 15, $sortField = null, $sortDirection = null, $dqlFilter = null)
     {
@@ -682,7 +683,7 @@ trait AdminControllerTrait
             'searchable_fields' => $searchableFields,
         ]);
 
-        return $this->get('easyadmin.paginator')->create($queryBuilder, $page, $maxPerPage);
+        return new QueryPaginator($queryBuilder, $page, $maxPerPage);
     }
 
     /**
